+++ bin-src/git/git.js
@@ -98,8 +98,11 @@ export async function getSlug() {
 // adhoc builds.
 
 // We could cache this, but it's probably pretty quick
-export async function getCommit() {
-  const result = await execGitCommand(`git --no-pager log -n 1 --format="%H ## %ct ## %ce ## %cn"`);
+export async function getCommit(revision = '') {
+  const result = await execGitCommand(
+    // Technically this yields the author info, not committer info
+    `git --no-pager log -n 1 --format="%H ## %ct ## %ae ## %an" ${revision}`
+  );
   const [commit, committedAtSeconds, committerEmail, committerName] = result.split(' ## ');
   return { commit, committedAt: committedAtSeconds * 1000, committerEmail, committerName };
 }