+++ bin-src/lib/getDependentStoryFiles.test.js
@@ -5,7 +5,10 @@ import { getRepositoryRoot } from '../git/git';
 import { getWorkingDir } from './utils';
 
 jest.mock('../git/git');
-jest.mock('./utils');
+jest.mock('./utils', () => {
+  const utils = jest.requireActual('./utils');
+  return { __esModule: true, ...utils, getWorkingDir: jest.fn() };
+});
 
 const CSF_GLOB = './src sync ^\\.\\/(?:(?!\\.)(?=.)[^/]*?\\.stories\\.js)$';
 const statsPath = 'preview-stats.json';
@@ -41,7 +44,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      './src/foo.stories.js': 'src/foo.stories.js',
+      './src/foo.stories.js': ['src/foo.stories.js'],
     });
   });
 
@@ -62,7 +65,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      './src/foo.stories.js': 'src/foo.stories.js',
+      './src/foo.stories.js': ['src/foo.stories.js'],
     });
   });
 
@@ -83,7 +86,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      './src/foo.stories.js': 'src/foo.stories.js',
+      './src/foo.stories.js': ['src/foo.stories.js'],
     });
   });
 
@@ -109,7 +112,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      './src/foo.stories.js': 'src/foo.stories.js',
+      './src/foo.stories.js': ['src/foo.stories.js'],
     });
   });
 
@@ -131,7 +134,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      './src/foo.stories.js': 'src/foo.stories.js',
+      './src/foo.stories.js': ['src/foo.stories.js', 'src/foo.js'],
     });
   });
 
@@ -158,7 +161,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      './src/foo.stories.js': 'services/webapp/src/foo.stories.js',
+      './src/foo.stories.js': ['services/webapp/src/foo.stories.js'],
     });
   });
 
@@ -187,7 +190,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      '/path/to/project/src/foo.stories.js': 'src/foo.stories.js',
+      '/path/to/project/src/foo.stories.js': ['src/foo.stories.js'],
     });
   });
 
@@ -218,7 +221,7 @@ describe('getDependentStoryFiles', () => {
     const ctx = getContext();
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual({
-      '/path/to/project/packages/webapp/src/foo.stories.js': 'packages/webapp/src/foo.stories.js',
+      '/path/to/project/packages/webapp/src/foo.stories.js': ['packages/webapp/src/foo.stories.js'],
     });
   });
 
@@ -255,7 +258,7 @@ describe('getDependentStoryFiles', () => {
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual(null);
     expect(ctx.turboSnap.bailReason).toEqual({
-      changedPackageFile: 'src/package.json',
+      changedPackageFiles: ['src/package.json'],
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(chalk`Found a package file change in {bold src/package.json}`)
@@ -280,7 +283,7 @@ describe('getDependentStoryFiles', () => {
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual(null);
     expect(ctx.turboSnap.bailReason).toEqual({
-      changedStorybookFile: 'path/to/storybook-config/file.js',
+      changedStorybookFiles: ['path/to/storybook-config/file.js'],
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(
@@ -307,7 +310,7 @@ describe('getDependentStoryFiles', () => {
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual(null);
     expect(ctx.turboSnap.bailReason).toEqual({
-      changedStorybookFile: 'path/to/storybook-config/file.js',
+      changedStorybookFiles: ['path/to/storybook-config/file.js'],
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(
@@ -335,7 +338,7 @@ describe('getDependentStoryFiles', () => {
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual(null);
     expect(ctx.turboSnap.bailReason).toEqual({
-      changedStorybookFile: 'path/to/storybook-config/file.js',
+      changedStorybookFiles: ['path/to/storybook-config/file.js', 'src/styles.js'],
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(
@@ -362,11 +365,43 @@ describe('getDependentStoryFiles', () => {
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual(null);
     expect(ctx.turboSnap.bailReason).toEqual({
-      changedStaticFile: 'path/to/statics/image.png',
+      changedStaticFiles: ['path/to/statics/image.png'],
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(chalk`Found a static file change in {bold path/to/statics/image.png}`)
     );
+  });
+
+  it('ignores untraced files', async () => {
+    const changedFiles = ['src/utils.js'];
+    const modules = [
+      {
+        id: 1,
+        name: './src/utils.js', // changed
+        reasons: [{ moduleName: './src/foo.js' }],
+      },
+      {
+        id: 2,
+        name: './src/foo.js', // untraced
+        reasons: [{ moduleName: './src/foo.stories.js' }],
+      },
+      {
+        id: 3,
+        name: './src/foo.stories.js',
+        reasons: [{ moduleName: CSF_GLOB }],
+      },
+      {
+        id: 999,
+        name: CSF_GLOB,
+        reasons: [{ moduleName: './.storybook/generated-stories-entry.js' }],
+      },
+    ];
+    const ctx = getContext({
+      storybook: { staticDir: ['path/to/statics'] },
+      options: { untraced: ['**/foo.js'] },
+    });
+    const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
+    expect(res).toEqual({});
   });
 });