+++ bin-src/lib/getOptions.js
@@ -1,6 +1,7 @@
 import path from 'path';
 import { parse } from 'url';
 
+import dependentOption from '../ui/messages/errors/dependentOption';
 import duplicatePatchBuild from '../ui/messages/errors/duplicatePatchBuild';
 import incompatibleOptions from '../ui/messages/errors/incompatibleOptions';
 import invalidExitOnceUploaded from '../ui/messages/errors/invalidExitOnceUploaded';
@@ -18,11 +19,16 @@ import inferredOptions from '../ui/messages/info/inferredOptions';
 import getStorybookConfiguration from './getStorybookConfiguration';
 
 const takeLast = (input) => (Array.isArray(input) ? input[input.length - 1] : input);
+const ensureArray = (input) => (Array.isArray(input) ? input : [input]);
 
 const resolveHomeDir = (filepath) =>
   filepath && filepath.startsWith('~') ? path.join(process.env.HOME, filepath.slice(1)) : filepath;
 
 const trueIfSet = (value) => (value === '' ? true : value);
+const undefinedIfEmpty = (array) => {
+  const filtered = array.filter(Boolean);
+  return filtered.length ? filtered : undefined;
+};
 
 export default async function getOptions({ argv, env, flags, log, packageJson }) {
   const fromCI = !!flags.ci || !!process.env.CI;
@@ -34,11 +40,12 @@ export default async function getOptions({ argv, env, flags, log, packageJson })
 
     only: flags.only,
     onlyChanged: trueIfSet(flags.onlyChanged),
-    externals: flags.externals,
+    untraced: undefinedIfEmpty(ensureArray(flags.untraced)),
+    externals: undefinedIfEmpty(ensureArray(flags.externals)),
+    traceChanged: trueIfSet(flags.traceChanged),
     list: flags.list,
     fromCI,
     skip: trueIfSet(flags.skip),
-    diagnostics: !!flags.diagnostics,
     dryRun: !!flags.dryRun,
     verbose: !!flags.debug,
     interactive: !flags.debug && !fromCI && !!flags.interactive && !!process.stdout.isTTY,
@@ -65,6 +72,7 @@ export default async function getOptions({ argv, env, flags, log, packageJson })
     port: flags.storybookPort,
     storybookBuildDir: takeLast(flags.storybookBuildDir),
     storybookBaseDir: flags.storybookBaseDir,
+    storybookConfigDir: flags.storybookConfigDir,
     storybookUrl: flags.storybookUrl,
     createTunnel: !flags.storybookUrl && env.CHROMATIC_CREATE_TUNNEL !== 'false',
 
@@ -120,6 +128,18 @@ export default async function getOptions({ argv, env, flags, log, packageJson })
 
   if (options.only && options.onlyChanged) {
     throw new Error(invalidSingularOptions(['--only', '--only-changed']));
+  }
+
+  if (options.untraced && !options.onlyChanged) {
+    throw new Error(dependentOption('--untraced', '--only-changed'));
+  }
+
+  if (options.externals && !options.onlyChanged) {
+    throw new Error(dependentOption('--externals', '--only-changed'));
+  }
+
+  if (options.traceChanged && !options.onlyChanged) {
+    throw new Error(dependentOption('--trace-changed', '--only-changed'));
   }
 
   // No need to start or build Storybook if we're going to fetch from a URL