+++ bin-src/ui/messages/errors/dependentOption.stories.js
@@ -1,0 +1,7 @@
+import dependentOption from './dependentOption';
+
+export default {
+  title: 'CLI/Messages/Errors',
+};
+
+export const DependentOption = () => dependentOption('--untraced', '--only-changed');