+++ bin-src/ui/messages/errors/missingGitHubInfo.js
@@ -7,7 +7,6 @@ import link from '../../components/link';
 export default ({ GITHUB_EVENT_NAME }) =>
   dedent(chalk`
     ${error} {bold Missing GitHub environment variable}
-    \`GITHUB_EVENT_NAME\` environment variable set to '${GITHUB_EVENT_NAME}', but
-    \`GITHUB_SHA\` and \`GITHUB_REF\` are not both set.
+    \`GITHUB_EVENT_NAME\` environment variable is set to '${GITHUB_EVENT_NAME}', but \`GITHUB_SHA\` and \`GITHUB_HEAD_REF\` are not both set.
     ${info} Read more at ${link('https://www.chromatic.com/docs/github-actions')}
   `);