+++ bin-src/ui/tasks/upload.js
@@ -46,11 +46,21 @@ export const tracing = (ctx) => {
   };
 };
 
-export const bailed = (ctx) => ({
-  status: 'pending',
-  title: 'TurboSnap disabled',
-  output: `Found a change in ${ctx.turboSnap.bailReason.changedFile}`,
-});
+export const bailed = (ctx) => {
+  const { changedPackageFiles, changedStorybookFiles, changedStaticFiles } =
+    ctx.turboSnap.bailReason;
+  const [firstFile, ...otherFiles] =
+    changedPackageFiles || changedStorybookFiles || changedStaticFiles;
+  const siblings = pluralize('sibling', otherFiles.length, true);
+  let output = `Found a change in ${firstFile}`;
+  if (otherFiles.length === 1) output += ' or its sibling';
+  if (otherFiles.length > 1) output += ` or one of its ${siblings}`;
+  return {
+    status: 'pending',
+    title: 'TurboSnap disabled',
+    output,
+  };
+};
 
 export const traced = (ctx) => {
   const files = pluralize('story file', Object.keys(ctx.onlyStoryFiles).length, true);